# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Evidence-Grounded Research Workbench API
  version: 0.2.0
  description: Full MVP API contract for the evidence-grounded multi-model long-horizon research workbench.
servers:
  - url: http://localhost:4000
    description: Local Fastify API
...
paths:
...
  /projects/{projectId}/claims:
    get:
      summary: List claims
      tags: [Claims]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Claims
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimListResponse'
  /claims/{claimId}:
    get:
      summary: Get claim by ID
      tags: [Claims]
      parameters:
        - $ref: '#/components/parameters/ClaimId'
      responses:
        '200':
          description: Claim
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update claim metadata or status
      tags: [Claims]
      parameters:
        - $ref: '#/components/parameters/ClaimId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClaimRequest'
      responses:
        '200':
          description: Updated claim
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimResponse'
...
  /claims/{claimId}/search-counter-evidence:
    post:
      summary: Search counter-evidence for claim
      tags: [Evidence]
      parameters:
        - $ref: '#/components/parameters/ClaimId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchEvidenceRequest'
      responses:
        '201':
          description: Counter-evidence search completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceListResponse'
...
  /projects/{projectId}/claims/{claimId}/confidence-history:
    get:
      tags: [Claims]
      summary: Get confidence history for a claim across deliberation rounds
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: claimId
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Confidence history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClaimConfidenceHistory'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/claims/{claimId}/dependencies:
    get:
      tags: [Claims]
      summary: Get claim dependencies (supports/depends-on relationships)
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: claimId
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Claim dependencies
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClaimDependency'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Claims]
      summary: Create a dependency between two claims
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: claimId
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetClaimId, relation]
              properties:
                targetClaimId: { type: string }
                relation:
                  type: string
                  enum: [supports, depends_on, contradicts, refines]
      responses:
        '201':
          description: Dependency created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimDependencyResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /projects/{projectId}/claims/{claimId}/dependencies/{dependencyId}:
    delete:
      tags: [Claims]
      summary: Remove a claim dependency
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: claimId
          in: path
          required: true
          schema: { type: string }
        - name: dependencyId
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dependency removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      deleted: { type: boolean }
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/claims/auto-detect-dependencies:
    post:
      tags: [Claims]
      summary: Auto-detect dependencies between claims using AI
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Auto-detection complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      created: { type: integer }
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Evidence Extras ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
...
    ClaimId:
      name: claimId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    Claim:
      type: object
      required: [id, projectId, ideaVersionId, text, type, requiresEvidence, criticality, status, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        ideaVersionId: { type: string }
        text: { type: string }
        type:
          type: string
          enum: [technical, product, market, business, legal, ux, research, risk, assumption]
        requiresEvidence: { type: boolean }
        criticality:
          type: string
          enum: [low, medium, high, blocking]
        status:
          type: string
          enum: [unverified, supported, partially_supported, contradicted, unsupported, needs_external_validation]
        confidence: { type: number, nullable: true }
        createdAt: { type: string, format: date-time }
    Evidence:
      type: object
      required: [id, projectId, title, sourceType, reliability, relevance, status, stalenessRisk, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        claimId: { type: string, nullable: true }
        discoveredByModelId: { type: string, nullable: true }
        sourceUrl: { type: string, nullable: true }
        title: { type: string }
        publisher: { type: string, nullable: true }
        publishedAt: { type: string, format: date-time, nullable: true }
        retrievedAt: { type: string, format: date-time }
        sourceType:
          type: string
          enum: [official, academic, government, company, news, benchmark, blog, forum, user_input, unknown]
        excerpt: { type: string, nullable: true }
        summary: { type: string, nullable: true }
        rawContentRef: { type: string, nullable: true }
        reliability:
          type: string
          enum: [pending, high, medium, low, unusable]
        relevance:
          type: string
          enum: [pending, direct, indirect, weak, irrelevant]
        status:
          type: string
          enum: [pending_review, accepted, accepted_with_caution, rejected, irrelevant, needs_better_source]
        stalenessRisk:
          type: string
          enum: [low, medium, high]
        createdAt: { type: string, format: date-time }
...
    UpdateClaimRequest:
      type: object
      properties:
        text: { type: string }
        type:
          type: string
          enum: [technical, product, market, business, legal, ux, research, risk, assumption]
        requiresEvidence: { type: boolean }
        criticality:
          type: string
          enum: [low, medium, high, blocking]
        status:
          type: string
          enum: [unverified, supported, partially_supported, contradicted, unsupported, needs_external_validation]
        confidence: { type: number }
...
    SearchEvidenceRequest:
      type: object
      properties:
        query: { type: string, nullable: true }
        provider:
          type: string
          enum: [manual, web, searxng]
          default: web
        maxResults:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
...
    ClaimResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Claim'
    ClaimListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Claim'
...
    EvidenceListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Evidence'
...